perf(minipay): stop shipping Privy, x402 and @solana/kit to clients that never use them
MiniPay clients get the injected connector and never mount PrivyTree, but
they were downloading, parsing and executing the whole Privy dependency
graph anyway — including x402 and @solana/kit — before rendering a
placeholder button.

wallet-provider.tsx already splits the Privy tree behind next/dynamic.
That split was bypassed by connect-button.tsx, which TopBar mounts on
every page including the map, and which reached @privy-io/react-auth
through two static imports: the useConnectWallet hook, and
PrivyReadyContext by way of wallet-provider-privy. Webpack resolves
static imports at build time, so the runtime PrivyReadyContext guard
could not prevent it.

Move PrivyReadyContext into its own React-only module, and move the
Privy-dependent half of ConnectButton behind next/dynamic with the
existing placeholder as its loading fallback. The shared button styles
move alongside so the placeholder and the real button cannot drift.

First Load JS on / drops 860 kB -> 339 kB. Raw chunk bytes for the route
drop 3043 kB -> 1304 kB, and logical assignment operators (||=, ??=,
Chrome 85+) fall from 21 to 6. No chunk on any route now carries Privy,
x402 or Solana code.

That last count is why this touches #196: the snippet reported as
throwing on a WebView 80 device traces to @solana/offchain-messages,
which is now off the path entirely. It does not close that issue — viem,
@noble/hashes and @wagmi/connectors still ship syntax above the Chrome 80
support floor, and all three are genuinely required.

Adds a source-level guard so the import cannot come back unnoticed; it
fails on exactly the reintroduced import that caused this.

Closes #208. Refs #196.
