+++ apps/web/src/components/connect-button.tsx
@@ -1,25 +1,9 @@
 "use client";
 
-import { useConnectWallet } from "@privy-io/react-auth";
-import { useAccount, useDisconnect } from "wagmi";
-import { useContext, useEffect, useRef, useState } from "react";
-import Link from "next/link";
-import { generateUsername } from "@/lib/username";
-import { useProfile } from "@/hooks/useProfile";
-import { useMaps } from "@/hooks/useMaps";
-import { PrivyReadyContext } from "./wallet-provider-privy";
-
-const buttonClassName = "pixel-btn pixel-btn-sm font-display";
-const buttonStyle: React.CSSProperties = {
-  fontSize: 8,
-  letterSpacing: 1.5,
-  minWidth: 108,
-  padding: "0 10px",
-  justifyContent: "center",
-  maxWidth: 160,
-  overflow: "hidden",
-  textOverflow: "ellipsis",
-};
+import dynamic from "next/dynamic";
+import { useContext, useEffect, useState } from "react";
+import { PrivyReadyContext } from "./privy-ready-context";
+import { buttonClassName, buttonStyle } from "./connect-button-styles";
 
 function ConnectButtonPlaceholder() {
   return (
@@ -30,6 +14,18 @@ function ConnectButtonPlaceholder() {
     </div>
   );
 }
+
+// The Privy-dependent half is lazy, and this module must stay free of any
+// `@privy-io/*` import — direct or transitive — for that to mean anything.
+// `TopBar` mounts this on every page, so a static import here lands in the
+// shared chunk and ships `@privy-io/react-auth` + `x402` + `@solana/kit` to
+// MiniPay clients, which never render the interactive half at all. The
+// runtime guard below cannot prevent that on its own: webpack resolves
+// imports at build time and has no idea `privyReady` will be `false`.
+const ConnectButtonInteractive = dynamic(
+  () => import("./connect-button-interactive"),
+  { ssr: false, loading: () => <ConnectButtonPlaceholder /> },
+);
 
 // `useConnectWallet` requires `PrivyProvider` to be an ancestor at the
 // moment it runs. With the MiniPay-first WalletProvider, that's only
@@ -58,96 +54,4 @@ export function ConnectButton() {
   if (!privyReady) return <ConnectButtonPlaceholder />;
 
   return <ConnectButtonInteractive />;
-}
-
-function ConnectButtonInteractive() {
-  const { connectWallet } = useConnectWallet();
-  const { disconnect } = useDisconnect();
-  const { isConnected, address } = useAccount();
-  const { currentMapId } = useMaps();
-  const { name: onChainName } = useProfile(address, currentMapId);
-  const [menuOpen, setMenuOpen] = useState(false);
-  const wrapperRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!menuOpen) return;
-    function onDocClick(e: MouseEvent) {
-      if (wrapperRef.current && !wrapperRef.current.contains(e.target as Node)) {
-        setMenuOpen(false);
-      }
-    }
-    document.addEventListener("mousedown", onDocClick);
-    return () => document.removeEventListener("mousedown", onDocClick);
-  }, [menuOpen]);
-
-  const username =
-    isConnected && address ? onChainName || generateUsername(address) : null;
-
-  const label = isConnected ? username ?? "…" : "CONNECT";
-
-  const onClick = () => {
-    if (isConnected) setMenuOpen((o) => !o);
-    else connectWallet();
-  };
-
-  const itemStyle: React.CSSProperties = {
-    display: "block",
-    width: "100%",
-    padding: "10px 12px",
-    fontSize: 8,
-    fontFamily: "'Press Start 2P', monospace",
-    letterSpacing: 1.5,
-    color: "var(--text)",
-    textDecoration: "none",
-    background: "transparent",
-    border: "none",
-    textAlign: "left",
-    cursor: "pointer",
-    whiteSpace: "nowrap",
-  };
-
-  return (
-    <div ref={wrapperRef} style={{ position: "relative" }}>
-      <button onClick={onClick} className={buttonClassName} style={buttonStyle}>
-        {label}
-      </button>
-      {menuOpen && isConnected && (
-        <div
-          role="menu"
-          style={{
-            position: "absolute",
-            top: "100%",
-            right: 0,
-            marginTop: 6,
-            background: "var(--card-bg)",
-            border: "1px solid var(--border)",
-            borderRadius: 6,
-            minWidth: 140,
-            zIndex: 100,
-            boxShadow: "0 4px 12px rgba(0,0,0,0.25)",
-            overflow: "hidden",
-          }}
-        >
-          <Link
-            href="/profile"
-            role="menuitem"
-            onClick={() => setMenuOpen(false)}
-            style={{ ...itemStyle, borderBottom: "1px solid var(--border)" }}
-          >
-            PROFILE
-          </Link>
-          <button
-            role="menuitem"
-            onClick={() => {
-              disconnect();
-              setMenuOpen(false);
-            }}
-            style={itemStyle}
-          >
-            LOG OUT
-          </button>
-        </div>
-      )}
-    </div>
-  );
 }
